Finish implementation of ActionCable
Linked up client and server to allow the sending and receiving of
messages.

NOTE: ActionCable specs are passing, but the actual ability to
send/receive messages isn't actually working.  The app loads/opens, but
when you navigate to '..3000/chat' and try sending a message it doesn't
send.  The text remains in the box after you hit enter and no messages
appear as they should.  Clearly there is something awry here and further
investigation required to get this functionality working.  It would be
disappointing to 'complete' the course without having the app actually
send/receive messages!